+++ constance/__init__.py
@@ -1,6 +1,6 @@
 from django.utils.functional import LazyObject
 
-__version__ = '1.2'
+__version__ = '2.0.0'
 
 default_app_config = 'constance.apps.ConstanceConfig'
 
@@ -9,5 +9,6 @@ class LazyConfig(LazyObject):
     def _setup(self):
         from .base import Config
         self._wrapped = Config()
+
 
 config = LazyConfig()