+++ constance/apps.py
@@ -1,5 +1,4 @@
 from django.db.models import signals
-from django import VERSION
 from django.apps import AppConfig
 from django.utils.translation import ugettext_lazy as _
 
@@ -13,7 +12,7 @@ class ConstanceConfig(AppConfig):
         signals.post_migrate.connect(self.create_perm,
                                      dispatch_uid='constance.create_perm')
 
-    def create_perm(self, *args, **kwargs):
+    def create_perm(self, using=None, *args, **kwargs):
         """
         Creates a fake content type and permission
         to be able to check for permissions
@@ -22,13 +21,12 @@ class ConstanceConfig(AppConfig):
         from django.contrib.contenttypes.models import ContentType
 
         if ContentType._meta.installed and Permission._meta.installed:
-            extra = {} if VERSION >= (1, 8) else {'name': 'config'}
-            content_type, created = ContentType.objects.get_or_create(
+            content_type, created = ContentType.objects.using(using).get_or_create(
                 app_label='constance',
                 model='config',
-                **extra)
+            )
 
-            permission, created = Permission.objects.get_or_create(
+            permission, created = Permission.objects.using(using).get_or_create(
                 name='Can change config',
                 content_type=content_type,
                 codename='change_config')