+++ constance/backends/database/__init__.py
@@ -4,7 +4,7 @@ from django.core.exceptions import ImproperlyConfigured
 from django.db.models.signals import post_save
 
 from .. import Backend
-from ... import settings
+from ... import settings, signals, config
 
 
 class DatabaseBackend(Backend):
@@ -25,8 +25,8 @@ class DatabaseBackend(Backend):
             if isinstance(self._cache, LocMemCache):
                 raise ImproperlyConfigured(
                     "The CONSTANCE_DATABASE_CACHE_BACKEND setting refers to a "
-                    "subclass of Django's local-memory backend (%r). Please set "
-                    "it to a backend that supports cross-process caching."
+                    "subclass of Django's local-memory backend (%r). Please "
+                    "set it to a backend that supports cross-process caching."
                     % settings.DATABASE_CACHE_BACKEND)
         else:
             self._cache = None
@@ -45,15 +45,15 @@ class DatabaseBackend(Backend):
             return
         autofill_values = {}
         autofill_values[full_cachekey] = 1
-        for key, value in self.mget(settings.CONFIG.keys()):
+        for key, value in self.mget(settings.CONFIG):
             autofill_values[self.add_prefix(key)] = value
         self._cache.set_many(autofill_values, timeout=self._autofill_timeout)
 
     def mget(self, keys):
         if not keys:
             return
-        keys = dict((self.add_prefix(key), key) for key in keys)
-        stored = self._model._default_manager.filter(key__in=keys.keys())
+        keys = {self.add_prefix(key): key for key in keys}
+        stored = self._model._default_manager.filter(key__in=keys)
         for const in stored:
             yield keys[const.key], const.value
 
@@ -77,6 +77,7 @@ class DatabaseBackend(Backend):
         return value
 
     def set(self, key, value):
+        old_value = self.get(key)
         constance, created = self._model._default_manager.get_or_create(
             key=self.add_prefix(key), defaults={'value': value}
         )
@@ -86,10 +87,13 @@ class DatabaseBackend(Backend):
         if self._cache:
             self._cache.set(key, value)
 
+        signals.config_updated.send(
+            sender=config, key=key, old_value=old_value, new_value=value
+        )
+
     def clear(self, sender, instance, created, **kwargs):
         if self._cache and not created:
-            keys = [self.add_prefix(k)
-                    for k in settings.CONFIG.keys()]
+            keys = [self.add_prefix(k) for k in settings.CONFIG]
             keys.append(self.add_prefix(self._autofill_cachekey))
             self._cache.delete_many(keys)
             self.autofill()