+++ tests/urls.py
@@ -1,8 +1,8 @@
 from django.contrib import admin
 
-from django.conf.urls import url, include
+from django.conf.urls import url
 
 
 urlpatterns = [
-    url(r'^admin/', include(admin.site.urls)),
+    url(r'^admin/', admin.site.urls),
 ]